+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Requires Elixir 1.15 or higher.
 ### Changes
 
 - `IdempotencyPlug.ETSStore` now uses `:set` instead of `:ordered_set` for faster lookups
+- Cached responses now use the original response headers exactly, dropping any headers set by upstream plugs
+- Support `:cached_headers` option to set custom headers like `Idempotent-Replayed` for cached responses
 
 ## v0.2.1 (2023-04-28)